WIP: ci: Make sure `tests-pass` is always run
A GitHub action job is marked as success when it is skipped. The `needs` rule means that this job is skipped when one of the tests fails. Therefore, the required job `tests-pass` will never block a PR from merging.

> When a job fails, any jobs that depend on the failed job are skipped and do not report a failure. A pull request that requires the check may not be blocked. To use a required check on a job that depends on other jobs, use the `always()` conditional expression in addition to `needs`, see "Using jobs in a workflow."

Source: https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/collaborating-on-repositories-with-code-quality-features/troubleshooting-required-status-checks#handling-skipped-but-required-checks